requirements.txt
# === Web Framework ===
flask==3.0.0
gunicorn==21.2.0

# === Data & Finance ===
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
pykrx==1.0.45

# === AI/LLM ===
google-generativeai==0.8.0
openai==1.12.0

# === Web Utilities ===
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
aiohttp==3.9.1

# === Utilities ===
python-dotenv==1.0.0
tqdm==4.66.1

# === Visualization ===
plotly==5.18.0
